refactored booking routes
